+++ src/zLOG/EventLogger.py
@@ -17,8 +17,6 @@ A logging module which handles event messages.
 This uses Vinay Sajip's PEP 282 logging module.
 """
 
-__version__='$Revision$'[11:-2]
-
 import logging
 import time
 
@@ -32,6 +30,18 @@ CUSTOM_TRACE = 5 # Mapping for zLOG.TRACE
 logging.addLevelName("BLATHER", CUSTOM_BLATHER)
 logging.addLevelName("TRACE", CUSTOM_TRACE)
 
+try:    # pragma: no cover
+   # Python 3
+   Exception.with_traceback
+
+   def fmt_raise(error):
+       raise error[0](error[1]).with_traceback(error[2])
+
+except AttributeError:   # pragma: no cover
+   # Python 2
+   def fmt_raise(error):
+       return error[0], error[1], error[2]
+
 
 def log_write(subsystem, severity, summary, detail, error):
     level = (zlog_to_pep282_severity_cache_get(severity) or
@@ -42,19 +52,7 @@ def log_write(subsystem, severity, summary, detail, error):
         msg = "%s\n%s" % (msg, detail)
 
     logger = logging.getLogger(subsystem)
-
-
-    # Since the logging module of Python does not allow to pass a
-    # traceback triple, we need to fake the exception. (See also
-    # Collector #1234).
-
-    if isinstance(error, tuple):
-        try:
-            raise error[0], error[1], error[2]
-        except:
-            pass
-
-    logger.log(level, msg, exc_info=(error is not None))
+    logger.log(level, msg, exc_info=error)
 
 
 def severity_string(severity, mapping={