+++ CHANGES.rst
@@ -6,12 +6,15 @@ Unreleased
 
 - Add coverage testing.
 
+- Add Python 3 compatibility.
+
 - Declare currently-supported Python versions (2.6, 2.7), and test them
-  via ``tox``.
 
 - Normalize package structure (``README.rst``, ``CHANGES.rst``).  Synthesize
   package description from README.rst and CHANGES.rst.
 
+- Use nose for testing instead of zope.testrunner and test
+  for 100% test coverage
 
 2.12.0 (2012-08-30)
 -------------------